Workaround change in tifffile's memmap feature
In `tifffile` version `0.13.0`+, it appears that the `memmap` keyword to
`asarray` has been dropped. Instead one must use `out` and set it to
`"memmap"`. To handle this change while remaining compatible with the
old versions of `tifffile`, we try the new syntax and fallback to old
syntax if a `TypeError` is raised (happens when the keyword doesn't
exist).