Use `.pages` attribute with TiffFile instances
In `tifffile` 0.13.0+, `TiffFile` objects no longer provide things like
`__getitem__` and `__len__`. Instead users are encourage to access the
`pages` attribute of `TiffFile` objects to get at this information.
Fortunately this works just fine with `tifffile` 0.12.x and 0.13.0+. So
just switch to using `pages` of `TiffFile` objects to find out the
number of pages and get individual pages.